unread [nfc]: Move UnreadStreamItem inside specific module; add jsdoc
This is really part of the private view-model code for this component,
so its most natural home is inside the component's module.
