+++ src/unread/UnreadCards.js
@@ -4,7 +4,6 @@ import React from 'react';
 import type { Node } from 'react';
 import { SectionList } from 'react-native';
 
-import type { UnreadStreamItem } from '../types';
 import { useDispatch, useSelector } from '../react-redux';
 import { SearchEmptyState } from '../common';
 import PmConversationList from '../pm-conversations/PmConversationList';
@@ -13,6 +12,34 @@ import TopicItem from '../streams/TopicItem';
 import { streamNarrow, topicNarrow } from '../utils/narrow';
 import { getUnreadConversations, getUnreadStreamsAndTopicsSansMuted } from '../selectors';
 import { doNarrow } from '../actions';
+
+/**
+ * An item in the data prepared for this UI by its helper selectors.
+ *
+ * See `getUnreadStreamsAndTopicsSansMuted`, and its helper
+ * `getUnreadStreamsAndTopics`.
+ *
+ * The exact collection of data included here is just an assortment of what
+ * the UI in this file happens to need.
+ */
+export type UnreadStreamItem = {|
+  key: string,
+  streamId: number,
+  streamName: string,
+  unread: number,
+  color: string,
+  isMuted: boolean,
+  isPinned: boolean,
+  isPrivate: boolean,
+  isWebPublic: boolean | void,
+  data: Array<{|
+    key: string,
+    topic: string,
+    unread: number,
+    isMuted: boolean,
+    lastUnreadMsgId: number,
+  |}>,
+|};
 
 type Props = $ReadOnly<{||}>;
