+++ src/unread/unreadSelectors.js
@@ -1,7 +1,8 @@
 /* @flow strict-local */
 import { createSelector } from 'reselect';
 
-import type { Narrow, Selector, UnreadStreamItem } from '../types';
+import type { Narrow, Selector } from '../types';
+import type { UnreadStreamItem } from './UnreadCards';
 import { caseInsensitiveCompareFunc } from '../utils/misc';
 import { getMute, isTopicMuted } from '../mute/muteModel';
 import { getOwnUserId } from '../users/userSelectors';
